Make requested changes
