+++ package.json
@@ -3,7 +3,6 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "eslint-config-prettier": "4.1.0",
     "react": "16.8.5",
     "react-dom": "16.8.5"
   },
@@ -15,7 +14,8 @@
     "prettier": "1.16.4",
     "react-scripts": "2.1.8",
     "stylelint": "9.10.1",
-    "typescript": "3.3.4000"
+    "typescript": "3.3.4000",
+    "eslint-config-prettier": "4.1.0"
   },
   "scripts": {
     "dev": "react-scripts start",
