+++ public/index.html
@@ -10,7 +10,7 @@
     <meta name="theme-color" content="#000000" />
     <title>React App</title>
   </head>
-  <body id="container">
+  <body class="root-container">
     <noscript>You need to enable JavaScript to run this app.</noscript>
     <div id="root" class="inherit-size"></div>
   </body>
