+++ src/app/styles/header/SearchBar.css
@@ -1,0 +1,29 @@
+.search-bar-container {
+  padding-top: 10px;
+  display: inline-block;
+  vertical-align: top;
+  width: calc(100% - 200px);
+}
+
+.search-bar-container__search-bar-and-close {
+  height: 32px;
+  width: 50%;
+  margin: auto;
+  position: relative;
+}
+
+.search-bar-container__search-bar::-webkit-search-cancel-button {
+  margin-right: 3px;
+}
+
+.search-bar-container__search-bar {
+  height: 32px;
+  width: 100%;
+  padding-left: 16px;
+  box-shadow: inset 0 0 0 1px rgba(0, 0, 0, 0.2);
+  background-color: #efefef;
+  outline: none;
+  font-family: 'Yandex Sans Text';
+  font-size: 15px;
+  line-height: 2.13;
+}
